# Faster (compared to ==0) is_zero