--- a/test/unittests/test_int128.cpp
+++ b/test/unittests/test_int128.cpp
@@ -192,6 +192,15 @@ static_assert(uint128(13) - 17 == uint128{0xffffffffffffffff, 0xfffffffffffffffc
 
 static_assert(-a == (~a + 1), "");
 static_assert(+a == a, "");
+
+static_assert(a != 0, "");
+static_assert(0 != s, "");
+static_assert(!is_zero(a), "");
+static_assert(!is_zero(s), "");
+static_assert(a - a == 0, "");
+static_assert(0 == s - a - a, "");
+static_assert(is_zero(a - a), "");
+static_assert(is_zero(s - a - a), "");
 }  // namespace static_test_arith
 #endif
 
@@ -366,13 +375,13 @@ TEST(int128, from_string)
 {
     constexpr auto ka = from_string<uint128>("18446744073709551617");
     static_assert(ka == uint128{1, 1}, "");
-    const auto *const sa = "18446744073709551617";
+    const auto* const sa = "18446744073709551617";
     const auto a = from_string<uint128>(sa);
     EXPECT_EQ(a, uint128(1, 1));
 
     constexpr auto kb = from_string<uint128>("0x300aabbccddeeff99");
     static_assert(kb == uint128{3, 0xaabbccddeeff99}, "");
-    const auto *const sb = "0x300aabbccddeeff99";
+    const auto* const sb = "0x300aabbccddeeff99";
     EXPECT_EQ(from_string<uint128>(sb), uint128(3, 0xaabbccddeeff99));
 }