Benchmark results on my MacBook:
```
unary_op<bool, uint256, op_eq_0>                             21 ns         21 ns   33263638
unary_op<bool, uint256, is_zero>                              8 ns          8 ns   87683038
unary_op<bool, uint512, op_eq_0>                             42 ns         42 ns   16355140
unary_op<bool, uint512, is_zero>                             14 ns         14 ns   49587714
```