@chfast 2020-10-02
```
------------------------------------------------------------------------
Benchmark                                 Time           CPU Iterations
------------------------------------------------------------------------
unary_op<bool, uint256, op_eq_0>          1 ns          1 ns 1000000000
unary_op<bool, uint256, is_zero>          1 ns          1 ns 1000000000
unary_op<bool, uint512, op_eq_0>          1 ns          1 ns  561977649
unary_op<bool, uint512, is_zero>          1 ns          1 ns  673760901
```

You should mark the functions `is_eq_0` and `is_zero` with `[[clang::noinline]]`. Otherwise the benchmarking loop gets vectorized, and this is not what we care about in interpreter.