@chfast 2020-10-02
I'm rather skeptical, because in isolation the `is_eq_0()` and `is_zero()` have identical assembly:
https://godbolt.org/z/63zzTj