@chfast 2021-02-17
I realized the `operator bool` does exactly that. So `x == 0` can be done as `static_cast<bool>(x)`.